requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
cloudinary==1.39.0
elevenlabs==0.3.0
openai-whisper==20231117
python-multipart==0.0.9
aiohttp==3.9.3 